Initial import
This code has been developed by me outside of version control for a long
time.

I have not decided on a license and commercialization yet, so this is
confidential for now.

All of this code is my own except for the referenced commons-codec
library, which I intend to eliminate as a dependency.