GPKeyInfoTemplate: allow for same id with multiple versions
The sysmoUSIM card emits a KIT that contains several versions of the
same keys. While we do not believe that this is what GP intends, we
allow for the key version to be set on the key and will ignore other
versions.

The code involved in checking keysets is a becoming a bit murky, but
that should be okay since this is just a pragmatic protection feature.

More thought might have to be put here for security purposes.